Enable syntax highlighting on python code.
By changing .python to .language-python for the Python code
blocks, we enable syntax highlighting. Currently, the syntax
highlighting doesn't work -- but if you merge in the lastest
code from the swcarpentry/styles repo -- then you will see
the syntax highlighting.

Currently, this is set up in preparation for this feature
being added.